requirements.txt
requests
django
djangorestframework
django-cors-headers
Pillow
drf-yasg
django-modeltranslation